Implement the FIDO AppID extension
This patch implements the appid extension behavior as described in
section 10.1 of the webauthn specification. See
https://www.w3.org/TR/webauthn/#sctn-appid-extension for details.
Namely, this will use the appid as presented in the extensions.appid
field for the ClientData.origin value, rather than the rpId. The
justification for this change is due to compatibility issues with
existing sites which use backward-compatible challenges even if the
authenticator device was never registered as a U2F key. This change has
been tested against Duo security for security keys enrolled with and
without U2F support.

This also modifies the ClientDataJSON generation algorithm to use
separators ':', and ',', without whitespace, to satisfy the requirements
of section 5.8.1.2 of the spec (Limited Verification Algorithm), which
states that the expected JSON structure should not have spaces. See
https://www.w3.org/TR/webauthn/#clientdatajson-verification for further
reading. In making this change, I have also used an OrderedDict for the
call to ClientData.build as that guarantees ordering even in versions of
python where this was not intially guaranteed for the standard dict.